feat: remove all old stories, update to storybook v8, update hook to use react 19
