+++ src/stories/Lottie.stories.js
@@ -28,6 +28,13 @@ export default {
       },
   };
 
+  const centerStyle = {
+    display: 'block',
+    margin: '10px auto',
+    textAlign: 'center',
+  };
+  const elementStyle = { margin: '0px auto' };
+
   const ToggleLottie = () => {
     const [state, setState] = useState({
         isStopped: true,
@@ -37,13 +44,6 @@ export default {
         isLike: false,
       });
 
-    const centerStyle = {
-      display: 'block',
-      margin: '10px auto',
-      textAlign: 'center',
-      pointer: 'cursor',
-    };
-    const elementStyle = { margin: "0px auto" };
     const defaultOptions = { animationData: TwitterHeart, loop: false, autoplay: false };
     const { isStopped, isPaused, direction } = state;
 
@@ -81,12 +81,7 @@ export default {
     });
     const [startFrame, setStartFrame] = useState(0);
     const [endFrame, setEndFrame] = useState(50);
-    const centerStyle = {
-      display: 'block',
-      margin: '10px auto',
-      textAlign: 'center',
-    };
-    const elementStyle = { margin: "0px auto" };
+
     const defaultOptions = { animationData: animationDataA };
     const { isStopped, isPaused, speed, direction } = state;
 
@@ -162,12 +157,7 @@ export default {
 
   const LottieLoop = () => {
     const [isTransitioned, setIsTransitioned] = useState(false);
-    const centerStyle = {
-        display: 'block',
-        margin: '10px auto',
-        textAlign: 'center',
-    };
-    const elementStyle = { margin: "0px auto" };
+
     const defaultOptions = {
       animationData: !isTransitioned ? TwitterHeart : BeatingHeart,
       loop: true,
@@ -207,6 +197,37 @@ export default {
     )
   }
 
+  const LottieTransition = () => {
+    const [showLoopedAnimation, setShowLoopedAnimation] = useState(true);
+
+    const animationOptionsWithLoop = {
+      animationData: animationDataA,
+      loop: true,
+    };
+
+    const animationOptionsWithoutLoop = {
+      animationData: BeatingHeart,
+      loop: false,
+    };
+
+    const clickHandler = () => {
+      setShowLoopedAnimation((prevState) => !prevState);
+    }
+
+    return (
+      <div>
+        <Lottie
+          style={elementStyle}
+          options={showLoopedAnimation ? animationOptionsWithLoop : animationOptionsWithoutLoop}
+          height={200}
+          width={200}
+        />
+        <p style={centerStyle}>This animation is {showLoopedAnimation ? 'looped' : 'not looped'}</p>
+        <button style={centerStyle} onClick={clickHandler}>switch</button>
+      </div>
+    )
+  }
+
   export const Default = {};
 
   export const Toggle = {
@@ -224,3 +245,7 @@ export default {
   export const LottieLoopTransition = {
     render: () => <LottieLoop />
   };
+
+  export const TransitionOptions = {
+    render: () => <LottieTransition />
+  }
